refactor: uses DataSource for mesh detail view
Changes MeshOverviewView to use DataSource and to solely be responsible for fetching data. All further presentational purposes where moved into the new MeshDetails and the existing MeshCharts components.

Adds the new MeshDetails component for the presentational purposes of the mesh detail view. Its pretty much a copy of what used to be in MeshOverviewView.

Changes MeshCharts to receive `meshInsight` as a prop instead of fetching the data itself. The mesh detail view always loads that data anyway, so this removes an unnecessary request from this view.
