+++ src/app/meshes/sources.ts
@@ -1,22 +1,18 @@
 import { DataSourceResponse } from '@/app/application/services/data-source/DataSourcePool'
 import type KumaApi from '@/services/kuma-api/KumaApi'
-import type {
-  PaginatedApiListResponse as CollectionResponse,
-} from '@/types/api.d'
-import type {
-  Mesh,
-  MeshInsight,
-} from '@/types/index.d'
+import type { PaginatedApiListResponse as CollectionResponse } from '@/types/api.d'
+import type { Mesh, MeshInsight } from '@/types/index.d'
 
-type MeshParams = {
-  mesh: string
+type DetailParams = {
+  name: string
 }
+
 type PaginationParams = {
   size: number
   page: number
 }
 
-type Closeable = {close: () => void}
+type Closeable = { close: () => void }
 
 export type MeshSource = DataSourceResponse<Mesh>
 export type MeshCollection = CollectionResponse<Mesh>
@@ -26,21 +22,29 @@ export type MeshInsightSource = DataSourceResponse<MeshInsight>
 
 export const sources = (api: KumaApi) => {
   return {
-    '/meshes': async (params: MeshParams & PaginationParams, source: Closeable) => {
+    '/meshes': async (params: PaginationParams, source: Closeable) => {
       source.close()
+
+      const size = params.size
       const offset = params.size * (params.page - 1)
-      return api.getAllMeshes({
-        size: params.size,
-        offset,
-      })
+
+      return api.getAllMeshes({ size, offset })
     },
-    '/:mesh/mesh': (params: MeshParams, source: Closeable) => {
+
+    '/meshes/:name': (params: DetailParams, source: Closeable) => {
       source.close()
-      return api.getMesh({ name: params.mesh })
+
+      const name = params.name
+
+      return api.getMesh({ name })
     },
-    '/:mesh/insights': async (params: MeshParams, source: Closeable) => {
+
+    '/mesh-insights/:name': async (params: DetailParams, source: Closeable) => {
       source.close()
-      return api.getMeshInsights({ name: params.mesh })
+
+      const name = params.name
+
+      return api.getMeshInsights({ name })
     },
 
   }
